requirements.txt
torch
torchvision
streamlit
matplotlib
opencv-python
pillow
numpy